resumebot/requirements.txt
django
pandas
PyPDF2
google-generativeai
python-dotenv
python-docx

docx2pdf

jinja2
